+++ src/packages/animate/index.taro.ts
@@ -1,5 +1,5 @@
 import { Animate } from './animate.taro'
 
-export type { AnimateType, AnimateAction } from './type'
+export type { AnimateType, AnimateAction } from './types'
 export type { AnimateProps } from './animate.taro'
 export default Animate
