+++ src/packages/animate/index.ts
@@ -1,5 +1,5 @@
 import { Animate } from './animate'
 
-export type { AnimateType, AnimateAction } from './type'
+export type { AnimateType, AnimateAction } from './types'
 export type { AnimateProps } from './animate'
 export default Animate
