+++ src/packages/animate/types.ts
