+++ src/packages/fixednav/index.taro.ts
@@ -1,8 +1,6 @@
 import { FixedNav } from './fixednav.taro'
 
-export type {
-  FixedNavProps,
-  FixedNavDirection,
-  FixedNavPosition,
-} from './fixednav.taro'
+export type { FixedNavDirection, FixedNavPosition, FixedNavItem } from './types'
+
+export type { FixedNavProps } from './fixednav.taro'
 export default FixedNav
