+++ src/packages/fixednav/index.ts
@@ -1,8 +1,6 @@
 import { FixedNav } from './fixednav'
 
-export type {
-  FixedNavProps,
-  FixedNavDirection,
-  FixedNavPosition,
-} from './fixednav'
+export type { FixedNavDirection, FixedNavPosition, FixedNavItem } from './types'
+
+export type { FixedNavProps } from './fixednav'
 export default FixedNav
